# backend/cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Healthcare platform API with FHIR R4 interoperability",
        "title": "Healthcare API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@healthcare.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/encounters/{encounterFhirId}/medications": {
            "get": {
                "description": "Returns all medication requests/prescriptions for an encounter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medications"
                ],
                "summary": "List medications by encounter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encounter FHIR ID",
                        "name": "encounterFhirId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/medication.MedicationResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new medication request/prescription for an encounter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medications"
                ],
                "summary": "Create a medication",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encounter FHIR ID",
                        "name": "encounterFhirId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Medication data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/medication.CreateMedicationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/medication.CreateMedicationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/patients/{patientFhirId}/studies": {
            "get": {
                "description": "Returns all imaging studies associated with a patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "imaging"
                ],
                "summary": "List imaging studies for a patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient FHIR ID",
                        "name": "patientFhirId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/imaging.HTTPImagingStudyResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Uploads a DICOM imaging study for a patient with metadata via multipart form",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "imaging"
                ],
                "summary": "Upload DICOM study",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient FHIR ID",
                        "name": "patientFhirId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Study title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Modality (e.g., CT, MR, XR)",
                        "name": "modality",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "DICOM file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/imaging.HTTPImagingStudyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/studies/{studyId}": {
            "get": {
                "description": "Returns details of a specific imaging study including download URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "imaging"
                ],
                "summary": "Get imaging study details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Study UUID",
                        "name": "studyId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/imaging.HTTPImagingStudyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "imaging.HTTPImagingStudyResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "download_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "modality": {
                    "type": "string"
                },
                "patient_fhir_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "study_instance_uid": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "medication.CreateMedicationRequest": {
            "type": "object",
            "properties": {
                "dosage_instructions": {
                    "type": "string"
                },
                "medication_code": {
                    "type": "string"
                },
                "medication_name": {
                    "type": "string"
                },
                "patient_fhir_id": {
                    "type": "string"
                },
                "practitioner_fhir_id": {
                    "type": "string"
                }
            }
        },
        "medication.CreateMedicationResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "dosage_instructions": {
                    "type": "string"
                },
                "encounter_fhir_id": {
                    "type": "string"
                },
                "fhir_id": {
                    "type": "string"
                },
                "medication_code": {
                    "type": "string"
                },
                "medication_name": {
                    "type": "string"
                },
                "patient_fhir_id": {
                    "type": "string"
                },
                "practitioner_fhir_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "medication.MedicationResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "dosage_instructions": {
                    "type": "string"
                },
                "encounter_fhir_id": {
                    "type": "string"
                },
                "fhir_id": {
                    "type": "string"
                },
                "medication_code": {
                    "type": "string"
                },
                "medication_name": {
                    "type": "string"
                },
                "patient_fhir_id": {
                    "type": "string"
                },
                "practitioner_fhir_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}